Merges guilds and fixes a [createworldcrash / badspawnerrors on first time creation.
