--- Data/Decoration/Malas/dark tides quest.cfg
@@ -1,113 +1,0 @@
-# a magical teleporter
-DarkTidesTeleporter 0x1822
-2105 1307 -50
-2018 1226 -90
-1068 458 -90
-
-# Maabus coffin
-MaabusCoffin 0x0 (SpawnLocation=(2024, 1224, -90))
-2023 1223 -83
-
-# Crystal cave barrier
-CrystalCaveBarrier 0x3967
-1193 512 -90
-1194 512 -90
-1195 512 -90
-1196 512 -90
-1197 512 -90
-
-# Kronus scroll box
-KronusScrollBox 0xE80
-1185 451 -90
-
-# Vault of secrets barrier
-VaultOfSecretsBarrier 0x49E
-1073 450 -90
-1073 451 -90
-1073 452 -90
-1078 450 -90
-1078 451 -90
-1078 452 -90
-
-# metal chest
-MetalChest 0x09AB (Hue=0x499)
-1076 453 -84
-1076 449 -84
-
-# Vault of secrets sign
-LocalizedSign 0x0BD0 (Hue=0x482;LabelNumber=1060189)
-1079 454 -84
-
-# metal signpost
-Static 0x0B9E
-
-# crate
-MediumCrate 0x0E3E
-1185 450 -90
-1185 450 -87
-1200 438 -87
-1203 439 -90
-1185 454 -87
-
-# crate
-MediumCrate 0x0E3F
-1185 448 -90
-1186 439 -87
-1201 438 -87
-1194 447 -90
-1194 448 -90
-
-# metal chest
-MetalGoldenChest 0x0E40 (Hue=0x835)
-1185 447 -90
-1184 446 -90
-1184 443 -86
-
-# metal chest
-MetalGoldenChest 0x0E41 (Hue=0x835)
-1184 445 -90
-1188 437 -90
-1188 437 -86
-1189 437 -90
-1194 438 -90
-
-# metal chest
-MetalGoldenChest 0x0E40
-1184 444 -90
-
-# metal chest
-MetalGoldenChest 0x0E40 (Hue=0x849)
-1184 443 -90
-
-# bookcase
-FullBookcase 0x0A9A
-1185 440 -90
-
-# crate
-LargeCrate 0x0E3D
-1186 439 -90
-1200 438 -90
-1201 438 -90
-1202 439 -90
-1202 439 -87
-1185 454 -90
-1185 453 -90
-1186 453 -87
-1186 453 -90
-
-# metal chest
-MetalGoldenChest 0x0E41 (Hue=0x849)
-1188 437 -82
-1190 437 -90
-
-# metal chest
-MetalGoldenChest 0x0E41 (Hue=0x482)
-1189 437 -86
-
-# small fish
-Static 0x0DD9
-1192 438 -90
-
-# bookcase
-FullBookcase 0x0A97
-1197 437 -90
