--- Data/Decoration/Malas/emino's undertaking quest.cfg
@@ -1,96 +1,0 @@
-# moongate
-Static 0x1FD7 (Hue=0x3F1)
-375 802 1
-
-# ankh
-AnkhNorth 0x0004 (Hue=0xFA)
-397 801 4
-
-#Guardian barrier
-GuardianBarrier 0x3967
-404 1140 0
-405 1140 0
-406 1140 0
-407 1140 0
-408 1140 0
-409 1140 0
-
-# Green teleporters
-GreenNinjaQuestTeleporter 0x051C
-418 804 -1
-
-Static 0x051C (Hue=0x17E)
-412 1123 0
-
-# Blue teleporters
-BlueNinjaQuestTeleporter 0x051C
-423 805 -1
-
-Static 0x051C (Hue=0x2)
-411 1117 0
-
-# Red teleporter
-Static 0x051C (Hue=0x21)
-421 1092 0
-
-# White teleporters
-WhiteNinjaQuestTeleporter 0x051C
-392 802 -1
-
-Static 0x051C (Hue=0x47E)
-412 1086 0
-
-# wooden chest
-EminosKatanaChest 0x0E42
-393 989 -15
-
-# floor spikes
-SpikeTrap 0x119A
-418 1084 0
-
-# floor saw
-SawTrap 0x11AC
-421 1077 0
-
-# hole
-FireColumnTrap 0x1B71
-403 1073 0
-410 1050 0
-408 1044 0
-409 1044 0
-410 1044 0
-411 1044 0
-412 1044 0
-413 1044 0
-
-# haramaki-do
-Static 0x2674 (Hue=0x48C)
-409 1009 -13
-
-# war mace
-Static 0x1406 (Hue=0x47E)
-409 1008 -13
-
-# kabuto
-Static 0x236D (Hue=0x48C)
-409 1008 -13
-
-# chest
-OrnateWoodenChest 0x280D
-436 999 -15
-
-# kimono
-Static 0x2681
-396 1149 0
-
-# ninja pants
-Static 0x2651 (Hue=0x530)
-397 1146 1
-
-# ninja mask
-Static 0x26A4 (Hue=0x530)
-395 1146 0
-
-# ninja shirt
-Static 0x2668 (Hue=0x530)
-396 1146 0
