--- Data/Decoration/Malas/haochi's trials quest.cfg
@@ -1,42 +1,0 @@
-# Candle
-HonorCandle 0x1433 (Unlit; Unprotected)
-419 731 17
-420 731 17
-422 731 17
-421 731 17
-
-# stone pavers
-Static 0x051C (Hue=0xFA)
-389 704 -1
-426 780 -1
-331 769 -1
-355 745 -1
-330 706 -1
-
-# moongate
-Static 0x1FD7 (Hue=0xFA)
-391 777 -1
-
-# ankh
-AnkhNorth 0x0004 (Hue=0xFA)
-351 733 18
-388 724 13
-
-# brass sign
-# Welcome to your new home, Samurai.
-LocalizedSign 0x0BD1 (LabelNumber=1063295)
-369 777 -1
-
-# chest
-HaochisTreasureChest 0x2811
-424 704 8
-424 705 8
-424 706 8
-
-# katana
-Static 0x13FF (Hue=0x482)
-414 703 9
-
-# Haochi's katana generator
-HaochisKatanaGenerator 0x1B7B
-415 703 1
