--- Data/Decoration/Old/Felucca/RuinedMagincia.cfg
@@ -1,1 +1,0 @@
-
