--- Data/Decoration/Old/Ilshenar/gargoylecity.cfg
@@ -1,192 +1,0 @@
-# metal door
-MetalDoor 0x0675 (Facing=WestCW; Hue=0x835)
-856 677 -40
-
-# metal door
-MetalDoor 0x0677 (Facing=EastCCW; Hue=0x835)
-853 672 -40
-
-# metal door
-Static 0x067B (Facing=EastCW; Hue=0x835)
-856 606 -40
-859 611 -40
-
-# barred metal door
-BarredMetalDoor2 0x1FED (Facing=WestCW; Hue=0x835)
-892 674 -40
-
-# barred metal door
-BarredMetalDoor2 0x1FF5 (Facing=SouthCW; Hue=0x835)
-896 657 -40
-896 662 -40
-896 667 -40
-
-# metal chest
-MetalChest 0x09AB
-848 719 0
-850 719 0
-854 607 -40
-854 607 -35
-857 670 -40
-
-# metal chest
-MetalGoldenChest 0x0E40
-828 725 0
-844 717 0
-852 608 -40
-852 609 -40
-852 609 -36
-852 609 -32
-852 611 -40
-852 611 -36
-852 613 -40
-852 613 -36
-852 675 -40
-852 675 -36
-852 676 -40
-852 676 -36
-
-# metal chest
-MetalGoldenChest 0x0E41
-832 712 0
-856 670 -40
-856 670 -36
-856 670 -32
-858 670 -40
-858 670 -36
-
-# metal chest
-MetalChest 0x0E7C
-852 612 -40
-852 675 -32
-
-# skinning knife
-Static 0x0EC5
-864 583 -34
-
-# anvil
-AnvilSouthAddon 0x0FB0
-786 662 0
-786 671 0
-791 662 0
-791 671 0
-796 662 0
-796 671 0
-
-# spinning wheel
-SpinningWheelEastAddon 0x1019
-893 620 -40
-893 622 -40
-893 624 -40
-
-# upright loom
-LoomEastAddon 0x1060
-893 617 -40
-893 626 -40
-
-# lever
-Static 0x1093
-760 645 10
-
-# butcher knife
-Static 0x13F7
-864 582 -34
-
-# platemail arms
-Static 0x1417 (Hue=0x44E)
-768 631 5
-
-# map
-Static 0x14EB
-876 661 -34
-
-# bellows
-Static 0x1986
-786 659 0
-786 668 0
-791 659 0
-791 668 0
-796 659 0
-796 668 0
-
-# forge
-Static 0x198A (Light=Circle300)
-786 660 0
-786 669 0
-791 660 0
-791 669 0
-796 660 0
-796 669 0
-
-# forge
-Static 0x198E
-786 661 0
-786 670 0
-791 661 0
-791 670 0
-796 661 0
-796 670 0
-
-# bellows
-Static 0x1992
-786 665 0
-786 674 0
-791 665 0
-791 674 0
-796 665 0
-796 674 0
-
-# forge
-Static 0x1996
-786 664 0
-786 673 0
-791 664 0
-791 673 0
-796 664 0
-796 673 0
-
-# forge
-Static 0x199A
-786 663 0
-786 672 0
-791 663 0
-791 672 0
-796 663 0
-796 672 0
-
-# heater shield
-Static 0x1B76 (Hue=0x44E)
-766 630 6
-
-# heater shield
-Static 0x1B77 (Hue=0x44E)
-768 630 5
-
-# Bell of Courage
-Static 0x1C12
-846 601 -34
-846 682 -34
-
-# brain
-Static 0x1CF0 (Hue=0x96E)
-771 631 11
-
-# table leg
-Static 0x1E7D (Hue=0x451)
-766 631 6
-
-# barrel staves
-Static 0x1EB1 (Hue=0x44E)
-775 632 -1
-
-# barrel staves
-Static 0x1EB2 (Hue=0x44E)
-773 632 -1
-
-# barrel staves
-Static 0x1EB3 (Hue=0x44E)
-767 631 2
-
-# statue
-Static 0x1F18 (Hue=0x451)
-767 630 6
