--- Data/Decoration/Old/Ilshenar/golemsandjukas.cfg
@@ -1,135 +1,0 @@
-# metal door
-MetalDoor 0x067D (Facing=SouthCW)
-1089 707 -80
-1094 699 -60
-1094 708 -60
-1103 704 -80
-
-# metal door
-MetalDoor 0x067F (Facing=NorthCCW)
-1089 706 -80
-1094 707 -60
-1094 698 -60
-1103 703 -80
-
-# metal door
-MetalDoor2 0x06CD (Facing=SouthCW)
-1103 652 -80
-1111 652 -80
-
-# metal door
-MetalDoor2 0x06CF (Facing=NorthCCW)
-1103 651 -80
-1111 651 -80
-
-# wooden door
-StrongWoodDoor 0x06E5 (Facing=WestCW)
-1071 671 -80
-1083 632 -60
-1083 658 -80
-1083 671 -80
-1083 671 -60
-1096 671 -80
-1097 658 -80
-1098 632 -60
-
-# wooden door
-StrongWoodDoor 0x06E7 (Facing=EastCCW)
-1062 635 -80
-1062 635 -60
-1062 635 -40
-1062 667 -80
-1062 667 -60
-1062 667 -40
-1083 632 -80
-1083 645 -80
-1096 632 -80
-1097 645 -80
-1105 635 -60
-1105 668 -80
-1105 668 -60
-1105 668 -40
-1105 635 -80
-
-# wooden door
-StrongWoodDoor 0x06E9 (Facing=WestCCW)
-1071 634 -40
-1074 669 -40
-1090 634 -40
-1092 669 -40
-1092 679 -40
-
-# wooden door
-StrongWoodDoor 0x06EB (Facing=EastCW)
-1105 635 -40
-
-# wooden door
-StrongWoodDoor 0x06ED (Facing=SouthCW)
-1064 651 -80
-
-# wooden door
-StrongWoodDoor 0x06F1 (Facing=SouthCCW)
-1067 630 -80
-1067 630 -60
-1067 673 -80
-1067 673 -60
-1100 630 -80
-1100 630 -60
-1100 673 -80
-1100 673 -60
-1100 673 -40
-
-# wooden door
-StrongWoodDoor 0x06F3 (Facing=NorthCW)
-1103 643 -60
-1103 660 -60
-
-# iron gate
-IronGate 0x082C (Facing=SouthCW)
-1064 644 -40
-1064 659 -40
-1081 644 -40
-1081 659 -40
-1086 644 -40
-1086 659 -40
-1103 644 -40
-1103 659 -40
-
-# bookcase
-FullBookcase 0x0A97
-1092 696 -60
-
-# bookcase
-FullBookcase 0x0A98
-1086 703 -80
-
-# bookcase
-FullBookcase 0x0A99
-1081 707 -60
-1081 708 -80
-
-# bookcase
-FullBookcase 0x0A9A
-1081 699 -60
-1081 704 -80
-
-# bookcase
-FullBookcase 0x0A9B
-1082 703 -80
-1088 696 -60
-
-# bookcase
-FullBookcase 0x0A9C
-1081 706 -60
-
-# khabur's journal
-KaburJournal 0x0FF1
-1073 633 -74
-
-# fropoz's journal
-FropozJournal 0x0FF1
-1084 707 -74
-
-# translated gargoyle journal
-TranslatedGargoyleJournal 0x0FF2
-1089 708 -54
