--- Data/Decoration/Trammel/uzeraan turmoil quest.cfg
@@ -1,81 +1,0 @@
-# a magical teleporter
-UzeraanTurmoilTeleporter 0x1822
-3597 2582 0
-
-# metal chest
-DaemonBloodChest 0x9AB
-3456 2558 50
-
-# Schmendrick cave teleporter
-Static 0x1822
-5260 1841 0
-
-# fire
-Static 0x19AB
-5210 1817 0
-5211 1814 0
-5205 1818 0
-
-# sparkle
-Static 0x374A
-5208 1817 0
-
-# sparkle
-Static 0x377A
-5209 1815 0
-
-# sparkle
-Static 0x376A
-5208 1815 0
-
-# magical sparkles
-Static 0x1153
-5207 1816 0
-5260 1842 0
-5261 1841 0
-
-# death vortex
-Static 0x3789
-5208 1815 0
-
-# magical sparkles
-Static 0x1156
-5261 1841 0
-5261 1842 0
-
-# cannon south
-Cannon 0x0 (CannonDirection=South)
-3678 2768 33
-3675 2767 31
-3593 2810 50
-
-# cannon west
-Cannon 0x0 (CannonDirection=West)
-3681 2776 39
-3683 2779 44
-3659 2790 47
-3705 2832 29
-
-# cannon east
-Cannon 0x0 (CannonDirection=East)
-3662 2795 68
-3596 2807 50
-
-# cannon north
-Cannon 0x0 (CannonDirection=North)
-3593 2804 50
-3702 2837 33
-
-# cannon balls
-Static 0xE74
-3706 2833 24
-3595 2808 49
-3594 2808 50
-3663 2793 67
-3661 2791 53
-3662 2793 67
-3676 2768 32
-3676 2767 31
-3682 2777 41
-3683 2777 42
-3681 2777 40
